Crate p3_miden_fri

Miden-specific FRI implementation with configurable folding factor.

Modules

prover

verifier

Structs

CommitPhaseProofStep

FriParameters

A set of parameters defining a specific instance of the FRI protocol.

FriProof

HidingFriPcs

A hiding FRI PCS. Both MMCSs must also be hiding; this is not enforced at compile time so it’s the user’s responsibility to configure.

QueryProof

TwoAdicFriFolding

TwoAdicFriPcs

A polynomial commitment scheme using FRI to generate opening proofs.

Traits

FriFoldingStrategy

Whereas FriParameters encompasses parameters the end user can set, FriFoldingStrategy is set by the PCS calling FRI, and abstracts over implementation details of the PCS.

Functions

create_benchmark_fri_params

Creates a set of FriParameters suitable for benchmarking. These parameters represent typical settings used in production-like scenarios.

create_benchmark_fri_params_zk

Creates a set of FriParameters suitable for benchmarking with zk enabled. These parameters represent typical settings used in production-like scenarios.

create_recursive_miden_fri_params

Creates a set of FriParameters for 96-bit conjectured security suitable for miden VM constraints proving in recursive context.

create_regular_miden_fri_params

Creates a set of FriParameters for 96-bit conjectured security suitable for miden VM constraints proving in non-recursive context.

create_test_fri_params

Creates a minimal set of FriParameters for testing purposes. These parameters are designed to reduce computational cost during tests.

create_test_fri_params_zk

Creates a minimal set of FriParameters for testing purposes, with zk enabled. These parameters are designed to reduce computational cost during tests.

Type Aliases

CommitmentWithOpeningPoints

A joint commitment to a collection of matrices and their opening at a collection of points.

ProverDataWithOpeningPoints

The Prover Data associated to a commitment to a collection of matrices and a list of points to open each matrix at.

TwoAdicFriFoldingForMmcs