p3_miden_fri

Trait FriFoldingStrategy

pub trait FriFoldingStrategy<F: Field, EF: ExtensionField<F>> {
    type InputProof;
    type InputError: Debug;

    // Required methods
    fn extra_query_index_bits(&self) -> usize;
    fn fold_row(
        &self,
        index: usize,
        log_height: usize,
        beta: EF,
        evals: impl Iterator<Item = EF>,
    ) -> EF;
    fn fold_matrix<M: Matrix<EF>>(&self, beta: EF, m: M) -> Vec<EF>;

    // Provided method
    fn log_folding_factor(&self) -> usize { ... }
}

Whereas FriParameters encompasses parameters the end user can set, FriFoldingStrategy is set by the PCS calling FRI, and abstracts over implementation details of the PCS.

Required Associated Types

type InputProof

type InputError: Debug

Required Methods

fn extra_query_index_bits(&self) -> usize

We can ask FRI to sample extra query bits (LSB) for our own purposes. They will be passed to our callbacks, but ignored (shifted off) by FRI.

fn fold_row( &self, index: usize, log_height: usize, beta: EF, evals: impl Iterator<Item = EF>, ) -> EF

Fold a row, returning a single column. Supporting arbitrary folding width that is a power of 2.

fn fold_matrix<M: Matrix<EF>>(&self, beta: EF, m: M) -> Vec<EF>

Same as applying fold_row to every row, possibly faster.

Provided Methods

fn log_folding_factor(&self) -> usize

Log of the folding factor (arity). Defaults to 1 (folding factor of 2).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.