p3_miden_fri::verifier

Function verify_fri

pub fn verify_fri<Folding, Val, Challenge, InputMmcs, FriMmcs, Challenger>(
    folding: &Folding,
    params: &FriParameters<FriMmcs>,
    proof: &FriProof<Challenge, FriMmcs, Challenger::Witness, Folding::InputProof>,
    challenger: &mut Challenger,
    commitments_with_opening_points: &[CommitmentWithOpeningPoints<Challenge, InputMmcs::Commitment, TwoAdicMultiplicativeCoset<Val>>],
    input_mmcs: &InputMmcs,
) -> Result<(), FriError<FriMmcs::Error, InputMmcs::Error>>
where
    Val: TwoAdicField,
    Challenge: ExtensionField<Val>,
    InputMmcs: Mmcs<Val>,
    FriMmcs: Mmcs<Challenge>,
    Challenger: FieldChallenger<Val> + GrindingChallenger + CanObserve<FriMmcs::Commitment>,
    Folding: FriFoldingStrategy<Val, Challenge, InputError = InputMmcs::Error, InputProof = Vec<BatchOpening<Val, InputMmcs>>>,

Verifies a FRI proof.

Arguments:

folding: The FRI folding scheme used by the prover.

params: The parameters for the specific FRI protocol instance.

proof: The proof to verify.

challenger: The Fiat-Shamir challenger.

commitments_with_opening_points: A vector of joint commitments to collections of matrices and openings of those matrices at a collection of points.